backend/requirements.txt
bcrypt==3.2.2
PyJWT==2.7.0
